@@ -1,14 +1,23 @@
 const gulp = require('gulp');
 
 // copy
-gulp.task('copy', done => {
+gulp.task('copy:esm', (done) => {
+  gulp.src([
+    'src/**/*.js',
+  ], {
+    base: './src/',
+  })
+    .pipe(gulp.dest('dist/esm'));
+
+  done();
+});
+
+gulp.task('copy:package', (done) => {
   gulp.src([
     'LICENSE',
-    'package-lock.json',
     'README.md',
   ])
     .pipe(gulp.dest('dist'));
 
   done();
 });
-
