@@ -6,7 +6,9 @@
  * @param as
  * @param type
  */
-export default ({ rel = 'preload', href, as, type }) => {
+export default ({
+  rel = 'preload', href, as, type,
+}) => {
   const preloadLink = document.createElement('link');
   preloadLink.rel = rel;
   preloadLink.href = href;
