@@ -1,13 +1,14 @@
+import entries from '../object/entries';
+
 /**
  * element add styles
  * @param element
  * @param styles(obj)
  */
 export default (element, styles) => {
-  for (const name in styles) {
-    // eslint-disable-next-line no-prototype-builtins
-    if (styles.hasOwnProperty(name)) {
-      element.style[name] = styles[name];
+  entries(styles).forEach(([k, v]) => {
+    if (Object.prototype.hasOwnProperty.call(styles, k)) {
+      element.style[k] = v;
     }
-  }
+  });
 };
