@@ -4,5 +4,5 @@
  * @param propertyName
  * @returns {string}
  */
-export default (element, propertyName) =>
-  window.getComputedStyle(element, null).getPropertyValue(propertyName);
+export default (element, propertyName) => window.getComputedStyle(element, null)
+  .getPropertyValue(propertyName);
