@@ -4,7 +4,4 @@
  * @param className
  * @returns {*}
  */
-export default (el, className = 'active') =>
-  Array.prototype.filter.call(el.parentNode.children, (child) =>
-    child.classList.contains(className)
-  );
+export default (el, className = 'active') => Array.prototype.filter.call(el.parentNode.children, (child) => child.classList.contains(className));
