@@ -8,14 +8,14 @@ export default (func, wait = 200, mustRun = 400) => {
   let timeout;
   let startTime = new Date();
 
-  return function () {
+  return (...__arguments__) => {
     const curTime = new Date();
 
     clearTimeout(timeout);
 
     if (curTime - startTime >= mustRun) {
       // If the specified trigger interval is reached, the handler is triggered
-      func.apply(this, arguments);
+      func.apply(this, __arguments__);
       startTime = curTime;
     } else {
       // Trigger interval not reached, reset timer
