@@ -3,5 +3,7 @@
  * @param video
  * @return {boolean}
  */
-export default (video) =>
-  video.currentTime > 0 && !video.paused && !video.ended && video.readyState > 2;
+export default (video) => video.currentTime > 0
+  && !video.paused
+  && !video.ended
+  && video.readyState > 2;
